fix: TLE & WA on data type

+++ 2024/B/main.cpp
@@ -1,7 +1,3 @@
-//
-// Created by cahInfor on 24/05/25.
-//
-
 #include <iostream>
 
 #include "jawaban.h"
@@ -9,7 +5,7 @@
 using namespace std;
 
 int main() {
-    int B, C, D;
+    long long B, C, D;
     cin >> B >> C >> D;
 
     cout << solve(B, C, D) << endl;
